backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic-settings==2.0.3
torch==2.1.1
torchaudio==2.1.1
transformers==4.35.2
librosa==0.10.1
numpy==1.24.3
scipy==1.11.4
pydub==0.25.1 